backend/pyproject.toml
[project]
name = "dietary-agent"
version = "3.1.0"
description = "Dietary Analysis Multi-Agent using Google ADK"
authors = [
    { name = "Saeed", email = "saeedagha@google.com" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "absl-py>=2.2.1",
    "google-adk>=0.0.2",
    "google-genai>=1.5.0",
    "pydantic>=2.9.2",
    "google-cloud-aiplatform[adk,agent-engines]>=1.93.0",
    "nest-asyncio>=1.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dietary_agent"]
